pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clerk-backend-api>=3.0.5",
    "fastapi>=0.116.1",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "svix>=1.69.0",
    "uvicorn>=0.35.0",
]
